requirements.txt
selenium
pandas
pyyaml


-e .